+++ HW_w2d1/SSD.h
@@ -10,8 +10,8 @@ public:
 	SSD(const SSD& obj);
 	~SSD();
 	char* GetName() const { return model; };
-	int GetPrice() const { return price; };
+	double GetPrice() const { return price; };
 	void SetName(char* n);
-	void SetPrice(int pr);
+	void SetPrice(double pr);
 	void Print();
 };
