+++ HW_w2d1/VPU.h
@@ -9,9 +9,9 @@ public:
 	VPU(const VPU& obj);
 	~VPU();
 	char* GetName() const { return model; };
-	int GetPrice() const { return price; };
+	double GetPrice() const { return price; };
 	void SetName(char* n);
-	void SetPrice(int pr);
+	void SetPrice(double pr);
 	void Print();
 };
